// src/app/contabil/models/obrigacoes/inputs/acessoriaMensalInput.ts
import { InputBase } from "../../../../shared/models";
import { TDiaTipo, TFeriadoTipo, TPeriodoMensal } from "../../enums";

export interface AcessoriaMensalInput extends InputBase {
    obrigacaoId?: number;
    competenciaInicial?: string;
    competenciaFinal?: string;
    vencimentoDia: number;
    vencimentoDiaTipo: TDiaTipo;
    prazoDiaAntesVencimento: number;
    periodoMensal: TPeriodoMensal;
    feriadoTipo: TFeriadoTipo;
}